UpdateReview.java with TotalSentiment()
The method totalSentiment works by finding the sentiment value of each word, making sure to exclude punctuation since it does not contribute to the overall value of it. We then add letters together until the method reaches a space, where then a method is called to find the sentiment value of the overall word found. The value is stored and added into sentiment and finally returned by totalSentiment